requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.0.1
streamlit==1.12.0
Pillow==9.0.1
requests==2.27.1